remove old ISA; abstract ISA as much as possible
