+++ Cargo.toml
@@ -21,28 +21,39 @@ name = "aluvm-stl"
 required-features = ["stl"]
 
 [dependencies]
-amplify = { version = "4.7.0", default-features = false, features = ["apfloat", "derive", "hex"] }
+amplify = { version = "4.7.0", default-features = false, features = ["derive", "hex"] }
 ascii-armor = { version = "0.7.2", optional = true }
 baid64 = "0.2.2"
 paste = "1"
 strict_encoding = { version = "2.7.0", default-features = false, features = ["float", "derive"] }
 strict_types = { version = "2.7.0", optional = true }
-sha2 = "0.10.8"
-blake3 = "1.5.4"
-ripemd = "0.1.3"
+sha2 = { version = "0.10.8", optional = true }
 secp256k1 = { version = "0.30.0", optional = true, features = ["global-context"] }
 curve25519-dalek = { version = "3.2.1", optional = true }
-half = "2.4.1" # Required to maintain MSRV
+half = { version = "2.4.1", optional = true } # Required to maintain MSRV
 serde_crate = { package = "serde", version = "1", optional = true }
 
 [features]
 default = ["std"]
-all = ["stl", "std", "log", "secp256k1", "curve25519", "serde", "ascii-armor"]
+all = [
+    "stl", "std", "log", "ascii-armor",
+    # Instruction set architecture extensions
+    "array", "str", "float", "sha", "secp256k1", "curve25519",
+    "serde"
+]
+
 stl = ["strict_types/armor", "std"]
 std = ["amplify/std"]
 log = ["std"]
 alloc = ["amplify/alloc"]
+
+# Instruction set architecture extensions
+array = []
+str = []
+float = ["amplify/apfloat", "half"]
+sha = ["sha2"]
 curve25519 = ["curve25519-dalek"]
+
 serde = ["serde_crate", "amplify/serde", "std", "strict_encoding/serde"]
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
